Refactor/remove custom MultipartFileSender to use Spring Boot Range requests
